backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
boto3>=1.35.0
pandas>=2.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pydantic==2.10.3
pydantic-core==2.27.1
email-validator>=2.1.0
starlette>=0.37.2
